requirements.txt
langchain-community==0.2.0
langchain-huggingface==0.1.0
langchain-openai==0.1.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
flask==3.0.0
werkzeug==3.0.1
openai==1.3.5
pandas==2.1.4
python-dotenv==1.0.0
